+++ src/main/java/io/hhplus/tdd/point/dto/PointChargeRequest.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.request;
+package io.hhplus.tdd.point.dto;
 
 public record PointChargeRequest(long amount) {
 }
