+++ src/main/java/io/hhplus/tdd/point/service/PointValidator.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.service;
 
 import org.springframework.stereotype.Component;
