+++ src/test/java/io/hhplus/tdd/point/PointServiceTest.java
@@ -2,8 +2,13 @@ package io.hhplus.tdd.point;
 
 import io.hhplus.tdd.database.PointHistoryTable;
 import io.hhplus.tdd.database.UserPointTable;
-import io.hhplus.tdd.point.request.PointChargeRequest;
-import io.hhplus.tdd.point.request.PointUseRequest;
+import io.hhplus.tdd.point.entity.PointHistory;
+import io.hhplus.tdd.point.entity.type.TransactionType;
+import io.hhplus.tdd.point.entity.UserPoint;
+import io.hhplus.tdd.point.dto.PointChargeRequest;
+import io.hhplus.tdd.point.dto.PointUseRequest;
+import io.hhplus.tdd.point.service.PointService;
+import io.hhplus.tdd.point.service.PointValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
